test: implement new test cases for activation module
